// src/app/models/student.ts
export class Student{
    constructor() {
      this.id = null;
      this.title = "";
      this.fio = "";
      this.category = "";
      this.country = "";
      this.text = "";
      this.year = "";
      this.city = "";
    }
    id: number;
    title: string;
    fio: string;
    category: string;
    country: string;
    text: string;
    year: string;
    city: string;  

  }